Moves CultureSeries to new Books namespace
