--- doc/unreleased/default/culture_series.md
@@ -1,15 +1,0 @@
-# Faker::CultureSeries
-
-```ruby
-Faker::CultureSeries.book #=> "The Player of Games"
-
-Faker::CultureSeries.culture_ship #=> "Fate Amenable To Change"
-
-Faker::CultureSeries.culture_ship_class #=> "General Systems Vehicle"
-
-Faker::CultureSeries.culture_ship_class_abv #=> "GSV"
-
-Faker::CultureSeries.civ #=> "Culture"
-
-Faker::CultureSeries.planet #=> "Xinth"
-```
