+++ unreleased_README.md
@@ -23,6 +23,7 @@ development.
 - [Generators](#generators)
     - [Default](#default)
     - [Blockchain](#blockchain)
+    - [Books](#books)
     - [Creature](#creature)
     - [Games](#games)
     - [Japanese Media](#japanese-media)
@@ -138,7 +139,6 @@ gem 'faker', :git => 'https://github.com/stympy/faker.git', :branch => 'master'
   - [Faker::Cosmere](doc/unreleased/default/cosmere.md)
   - [Faker::Crypto](doc/unreleased/default/crypto.md)
   - [Faker::CryptoCoin](doc/unreleased/default/crypto_coin.md)
-  - [Faker::CultureSeries](doc/unreleased/default/culture_series.md)
   - [Faker::Currency](doc/unreleased/default/currency.md)
   - [Faker::Date](doc/unreleased/default/date.md)
   - [Faker::DcComics](doc/unreleased/default/dc_comics.md)
@@ -214,6 +214,9 @@ gem 'faker', :git => 'https://github.com/stympy/faker.git', :branch => 'master'
   - [Faker::Blockchain::Bitcoin](doc/unreleased/blockchain/bitcoin.md)
   - [Faker::Blockchain::Ethereum](doc/unreleased/blockchain/ethereum.md)
   - [Faker::Blockchain::Tezos](doc/unreleased/blockchain/tezos.md)
+
+### Books
+  - [Faker::Books::CultureSeries](doc/unreleased/books/culture_series.md)
 
 ### Creature
   - [Faker::Creature::Animal](doc/unreleased/creature/animal.md)
